+++ src/pytezos/context/mixin.py
@@ -8,7 +8,7 @@ from pytezos.jupyter import InlineDocstring
 from pytezos.rpc import RpcMultiNode, RpcNode, ShellQuery
 from pytezos.rpc.errors import RpcError
 
-default_network = 'hangzhounet'
+default_network = 'ithacanet'
 default_key = 'edsk33N474hxzA4sKeWVM6iuGNGDpX2mGwHNxEA4UbWS8sW3Ta3NKH'  # please, use responsibly
 default_key_hash = 'tz1grSQDByRpnVs7sPtaprNZRp531ZKz6Jmm'
 
@@ -29,6 +29,7 @@ nodes = {
     'kaizen': ['https://rpc.tzkt.io/granadanet'],
     'kaizennet': ['https://rpc.tzkt.io/granadanet'],
     'hangzhounet': ['https://rpc.tzkt.io/hangzhou2net'],
+    'ithacanet': ['https://rpc.tzkt.io/ithacanet']
 }
 keys = {
     'alice': alice_key,